requirements.txt
pygame
opencv-python
requests
pydub
ffmpeg-python